requirements.txt
Django>=5.0.0
djangorestframework
django-cors-headers
psycopg2-binary
python-dotenv
Pillow
djangorestframework-simplejwt
gunicorn
whitenoise
